+++ src/animation/Animation.h
@@ -211,7 +211,7 @@ void stopAnimation(Entity * entity, size_t layer = 0);
 short ANIM_GetAltIdx(const ANIM_HANDLE & ah, long old);
 void ANIM_Set(AnimLayer & layer, ANIM_HANDLE * anim);
 
-Vec3f GetAnimTotalTranslate(ANIM_HANDLE * eanim, long alt_idx);
+Vec3f GetAnimTotalTranslate(ANIM_HANDLE * eanim, size_t alt_idx);
 
 void EERIE_ANIMMANAGER_ClearAll();
 void EERIE_ANIMMANAGER_PurgeUnused();
